Fix silly mistake creating fail2ban vacuum cron job
The current code creates a broken cron job. This fixes that.

It's also worth note that by default, TurnKey doesn't include sqlite,
so this cron job will never do anything. So in retrospect, it's not
required. I'll leave it be for now though as it will be good if users
do decide to install sqlite.